Now with cmake build files. No install targets or settings yet.
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/frontends/console/trunk@243 6c50cc09-817d-0410-90cf-7b22d5ef511c

+++ Glut/Window.cpp
@@ -352,7 +352,9 @@ void megamol::viewer::Window::glutReshapeCallback(int w, int h) {
             thisWindow()->height = h;
         }
         ::glViewport(0, 0, w, h);
-        unsigned int size[2] = { w, h };
+        unsigned int size[2] = {
+            static_cast<unsigned int>(w),
+            static_cast<unsigned int>(h) };
         thisWindow()->resizeCallback.Call(*thisWindow(), size);
 
     } catch(vislib::Exception ex) {
